+++ javascope/DataCached.java
@@ -5,10 +5,10 @@ public class DataCached implements Serializable
  {
     String expression;
     String experiment;
-    int    shot;
+    long    shot;
     Object data;
     
-    DataCached(String expression, String experiment, int shot, Object data)
+    DataCached(String expression, String experiment, long shot, Object data)
     {
         this.expression = expression;
         this.shot = shot;
@@ -16,7 +16,7 @@ public class DataCached implements Serializable
         this.data = data;
     }
     
-    public boolean equals(String expression, String experiment, int shot)
+    public boolean equals(String expression, String experiment, long shot)
     {
         return ( this.expression.equals(expression) && 
                  this.shot == shot &&
